providers:checkbox: Test definition fixes for stress.
Removed a bad dependency from an attachment test. If the depended-on
test fails, we will have no attachment and thus no way to debug
failures.

Fixed bad number of iterations in the network interface restart test.
Instructions say the test will run 100 times but this was not reflected
in the definition.