Modify rotation test to use DBUS directly
On some NVidia cards, calling xrandr script to rotate the screen
programatically may not work. Triggering screen rotation using DBUS
directly is closer to what is done when the user manually modifies
the rotation settings in the System Settings Display screen.

We add a new script to handle this and modify the graphics rotation
test job to use it.