provider:checkbox: Fix AMD hybrid testing graphics_env script
Previous version of graphics_env script was using the content of
gpu-manager.log to check how many cards were available. Unfortunately, this
file is not reliable because it uses udev which, as of today, does only care
about proprietary drivers (nvidia, fglrx).

We change the method to count the number of GPUs with something easier and
more reliable.