add job that checks APSTE availabilty on NVMe drives
I decided to go with device.driver == 'nvme' and device.category == 'OTHER'"
template-filter to make sure only one job is generated (per controller), and
that the template will be given nvmeN name (N being 0, 1..), so the check for
pm_qos file would be possible. If I'd narrow it down to category == 'DISK' then
the drives would be reported, like nvme0n1, which is fine for the nvme cli
tool, but it would require some obscure tweaking to get 'the parent' for the
sysfs check.