remove shutter.xml if already exists
Fixes: LP:1691724

There was a case of shutter.xml being owned by root - possibly due to running
the job as root. This patch is kind of work around. It removes the previously
created shutter.xml if it exists, making shutil.copy() not throw an exception.
In a situation where the parent dir is also root's xrandr_cycle will report
a warning and use existing shutter.xml

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>