pm_test: Unset LD_LIBRARY_PATH, PYTHONPATH and PYTHONHOME to start system terminal
gnome-terminal is using a python3 starter script that has to load the system
wide modules/library to work properly.
That's not the case in pm_test is started from a checkbox snap where all those
env var point to paths inside the snap.

This patch unset them to ensure gnome-terminal (via x-terminal-emulator) starts
with the right env.

For the record the error was:
ImportError: module encoding not found