bin: let PRIME offloading on AMD GPUs work on any Ubuntu release
When the script to set PRIME offloading on AMD GPUs was initially
writted, we were still certifying Trusty (14.04) configs, and since
Trusty didn't have this feature, we needed to bypass it.

Now that all the Ubuntu releases we certify support AMD GPU offloading
(i.e. DRI_PRIME=1 for amdgpu drivers), we get rid of the check on Ubuntu
release.

LP: #1831719